keep only the connection related codes in fastbuilder/core. this will be very helpful in future refactoring (allow them to share the same mc related core part)
