+++ fastbuilder/credentials/uesr_input.go
@@ -1,0 +1,41 @@
+package credentials
+
+import (
+	"bufio"
+	"fmt"
+	"os"
+	"strings"
+	"syscall"
+
+	I18n "phoenixbuilder/fastbuilder/i18n"
+
+	"github.com/pterm/pterm"
+	"golang.org/x/term"
+)
+
+func getInput() (string, error) {
+	reader := bufio.NewReader(os.Stdin)
+	inp, err := reader.ReadString('\n')
+	inpl := strings.TrimRight(inp, "\r\n")
+	return inpl, err
+}
+
+func GetInputUserName() (string, error) {
+	reader := bufio.NewReader(os.Stdin)
+	pterm.Printf(I18n.T(I18n.Enter_FBUC_Username))
+	fbusername, err := reader.ReadString('\n')
+	return fbusername, err
+}
+
+func GetRentalServerCode() (string, string, error) {
+	reader := bufio.NewReader(os.Stdin)
+	fmt.Printf(I18n.T(I18n.Enter_Rental_Server_Code))
+	code, err := reader.ReadString('\n')
+	if err != nil {
+		return "", "", err
+	}
+	fmt.Printf(I18n.T(I18n.Enter_Rental_Server_Password))
+	bytePassword, err := term.ReadPassword(int(syscall.Stdin))
+	fmt.Printf("\n")
+	return strings.TrimRight(code, "\r\n"), string(bytePassword), err
+}
